Feature: Add link to feed on subscription button press
I was still getting non-converts, after building the modal. This indicated
to me that people are expecting something still different from what they're
seeing. By opening a feed, this will give more advanced users the option
to copy the feed and use it in feedly, reader, or another feed handler.